% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{avg_resp_probs}
\alias{avg_resp_probs}
\title{Average Response Probabilities}
\usage{
avg_resp_probs(resp_probs_list)
}
\arguments{
\item{resp_probs_list}{List of response probability vectors}
}
\value{
Vector of average response probabilities
}
\description{
Function that takes a list of model response probabilities and averages.
}
