% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{diva_grid_search}
\alias{diva_grid_search}
\title{DIVA Grid Search}
\usage{
diva_grid_search(param_list, num_inits, input_list, fit_type,
  crit_fit_vector = NULL)
}
\arguments{
\item{param_list}{List of named parameters to be combined and evaluated for
the DIVA model}

\item{num_inits}{Scalar for number of random initializations to be averaged
across for the response probability calculation of each parameter combination.}

\item{input_list}{List of inputs and labels for the grid search
\itemize{
    \item \code{ins} Matrix of inputs for selected category structure,
        R (stimuli) x C (features)
    \item \code{labels} Vector of labels for selected category structure
        indexed to the input matrx
    }}

\item{fit_type}{Character specifying the type of fit that is desired.
\itemize{
    \item \code{'best'} for the most accurate or best fit
    \item \code{'crit'} for the closest match to a provided vector of
    response probabilities
    }}

\item{crit_fit_vector}{Vector of response probabilities for the
\code{'crit'} procedure.}
}
\value{
List consisting of models, response probabilities and best model result.
}
\description{
Runs a grid search over a set of provided parameters and produces averaged
response probabilities
}
