% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{diva_grid_search_par}
\alias{diva_grid_search_par}
\title{Parallelized DIVA Grid Search}
\usage{
diva_grid_search_par(param_list, num_inits, input_list, fit_type = NULL,
  crit_fit_vector = NULL, state = NULL, procs = NULL)
}
\arguments{
\item{param_list}{List of named parameters to be combined and evaluated for
the DIVA model}

\item{num_inits}{Scalar for number of random initializations to be averaged
across for the response probability calculation of each parameter combination.}

\item{input_list}{List of inputs and labels for the grid search
\itemize{
    \item \code{ins} Matrix of inputs for selected category structure,
        R (stimuli) x C (features)
    \item \code{labels} Vector of labels for selected category structure
        indexed to the input matrx
    }}

\item{fit_type}{Character specifying the type of fit that is desired.
\itemize{
    \item \code{'best'} for the most accurate fit
    \item \code{'crit'} for the closest match to a provided vector of
    response probabilities
    }}

\item{crit_fit_vector}{Vector of response probabilities for the
\code{'crit'} procedure.}

\item{state}{List of model parameters. Useful for setting parameters that
are not subject to the grid search. \code{generate_state} is used if
no state is provided. Use \code{?generate_state} to examine defaults.}

\item{procs}{Scalar for number of processors to use in parallelization.
Defaults to \code{detectCores() - 2}}
}
\value{
List consisting of models, response probabilities and best model result.
}
\description{
Runs a grid search over a set of provided parameters and produces averaged
response probabilities
}
