% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{generate_state}
\alias{generate_state}
\title{Construct DIVA state list}
\usage{
generate_state(input, categories, colskip, continuous, make_wts,
  wts_range = NULL, wts_center = NULL, num_hids = NULL,
  learning_rate = NULL, beta_val = NULL, phi = NULL, model_seed = NULL)
}
\arguments{
\item{input}{Complete matrix of inputs for training}

\item{categories}{Vector of category assignment values}

\item{colskip}{Scalar for number of columns to skip in the tr matrix}

\item{continuous}{Boolean value indicating if inputs are continuous}

\item{make_wts}{Boolean value indicating if initial weights should be generated}

\item{wts_range}{Scalar value for the range of the generated weights}

\item{wts_center}{Scalar value for the center of the weights}

\item{num_hids}{Scalar value for the number of hidden units in the model architecture}

\item{learning_rate}{Learning rate for weight updates through backpropagation}

\item{beta_val}{Scalar value for the beta parameter}

\item{phi}{Scalar value for response mapping (Default = 1, meaning no effect)}

\item{model_seed}{Scalar value to set the random seed}
}
\value{
List of the model hyperparameters and weights
}
\description{
Construct the state list. The primary use of this function is to construct a
state list with the default DIVA parameters or generate data-appropriate weights
(particularly when the random seed needs to be set).
}
