% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{get_test_inputs}
\alias{get_test_inputs}
\title{Produce model inputs from a set of classic category structures}
\usage{
get_test_inputs(target_cats)
}
\arguments{
\item{target_cats}{String corresponding to category structure label
\itemize{
    \item \code{unifr} Unidimensional rule plus family resemblance structure
    \item \code{unfr1}, \code{unfr2} UNI-FR (separated for autoencoder modeling)
    \item \code{fr4d} Four-dimensional family resemblance structure
    \item \code{mns81e1_ls} Medin and Schwanenflugel '81 Experiment 1 Linearly Separable Structure
    \item \code{mns81e1_nls} Medin and Schwanenflugel '81 Experiment 1 Non-Linearly Separable Structure
    \item \code{mns81e4_ls} Medin and Schwanenflugel '81 Experiment 4 Linearly Separable Structure
    \item \code{mns81e4_nls} Medin and Schwanenflugel '81 Experiment 4 Non-Linearly Separable Structure
    \item \code{type1}, \code{type2}, \code{typeN}... Shepard, Hovland and Jenkin's elemental types
    \item \code{multiclass} 4 class problem built from SHJ Type II
    }}
}
\value{
A list of inputs and labels
\itemize{
    \item \code{ins} Inputs for selected category structure
    \item \code{labels} Labels for selected category structure
    }
}
\description{
Function to grab inputs that might be useful for model testing
}
