% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{plot_training}
\alias{plot_training}
\title{Plot Training}
\usage{
plot_training(model_list, model_names = NULL)
}
\arguments{
\item{model_list}{List of model response probabilities across training blocks}

\item{model_names}{Optional vector of character values corresponding to model names}
}
\value{
Training line plot
}
\description{
Plots the training curve for N models.
}
