% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{response_probs}
\alias{response_probs}
\title{Response Probabilities}
\usage{
response_probs(tr, out_probs, blocks = TRUE)
}
\arguments{
\item{tr}{Matrix used to train the model.}

\item{out_probs}{Matrix of output probabilities produced by the model.}

\item{blocks}{Boolean to toggle block averaged classification probabilities, default is TRUE}
}
\value{
Vector of classification probabilities for the target class
}
\description{
Produces classification probability for the target class, by item or by block.
}
