% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slpDIVAdev.R
\name{slpDIVAdev}
\alias{slpDIVAdev}
\title{Train stateful list processor DIVA}
\usage{
slpDIVAdev(st, tr, xtdo = FALSE)
}
\arguments{
\item{st}{A list of the model parameters}

\item{tr}{A matrix of the input and class labels}

\item{xtdo}{A boolean value indicating if extended output is
desired}
}
\value{
List including a matrix of model classification
    probabilities and list of model's final state
}
\description{
Train stateful list processor DIVA
}
