% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{tr_add}
\alias{tr_add}
\title{Add trials to an initialized tr object}
\usage{
tr_add(inputs, tr, labels = NULL, blocks = 1, ctrl = NULL,
  shuffle = FALSE, reset = FALSE)
}
\arguments{
\item{inputs}{Matrix of feature values for each item}

\item{tr}{Initialized trial object}

\item{labels}{Integer class labels for each input. Default NULL}

\item{blocks}{Integer number of repetitions. Default 1}

\item{ctrl}{Integer control parameter, applying to all inputs. Default 2}

\item{shuffle}{Boolean, shuffle each block. Default FALSE}

\item{reset}{Boolean, reset state on first trial (ctrl=1). Default FALSE}
}
\value{
An updated dataframe
}
\description{
Add trials to an initialized tr object
}
