% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{tr_init}
\alias{tr_init}
\title{Initialize a tr object}
\usage{
tr_init(n_feats, n_cats, feature_type = "numeric")
}
\arguments{
\item{n_feats}{Number of features (integer, > 0)}

\item{feature_type}{String type: numeric (default), logical, etc}
}
\value{
An initialized dataframe with the appropriate columns
}
\description{
Initialize a tr object
}

