% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlearn_suppls.r
\name{get_test_inputs}
\alias{get_test_inputs}
\title{Function to grab inputs that might be useful for model testing}
\usage{
get_test_inputs(target_cats)
}
\arguments{
\item{target_cats}{Variable to choose a category: 
    unifr - (+ unfr1 and unfr2 for autoencoder)
    type# - Shepard, Hovland and Jenkin's elemental types, e.g., type2, type 3...
    multiclass - 4 class problem (shj type 2)
get_test_inputs()}
}
\description{
Function to grab inputs that might be useful for model testing
}

